Extract editor playback HLS warmup
